Add files via upload
Ready to finalize a POS cash transaction? This onscreen UI prompts the total due and provides a touch interface to pay with cash amounts. Instead of an input box and the use of a keyboard (or an onscreen keyboard), this alternative may be more functional for a touch only POS terminal.